AArch64: Deprecate -mabi=ilp32
ILP32 was originally intended to make porting to AArch64 easier.  Support was
never merged in the Linux kernel or GLIBC, so it has been unsupported for many
years.  There isn't a benefit in keeping unsupported features forever, so
deprecate it now (and it could be removed in a future release).

gcc:
	* config/aarch64/aarch64.cc (aarch64_override_options): Add warning.
	* doc/invoke.texi: Document -mabi=ilp32 as deprecated.

gcc/testsuite:
	* gcc.target/aarch64/inline-mem-set-pr112804.c: Add -Wno-deprecated.
	* gcc.target/aarch64/pr100518.c: Likewise.
	* gcc.target/aarch64/pr113114.c: Likewise.
	* gcc.target/aarch64/pr80295.c: Likewise.
	* gcc.target/aarch64/pr94201.c: Likewise.
	* gcc.target/aarch64/pr94577.c: Likewise.
	* gcc.target/aarch64/sve/pr108603.c: Likewise.